requirements.txt
crewai
fastapi
python-dotenv